requirements/base.txt
transformers==4.36.0
torch==2.1.0
gradio==4.44.1
huggingface-hub==0.19.4
datasets==2.15.0
numpy==1.24.3
pandas
scikit-learn==1.3.2
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
mlflow==2.9.0
wandb==0.16.1
prometheus-client==0.19.0
pyyaml==6.0.1
plotly==6.3.0
accelerate==0.25.0
fastapi[all]
